+++ apps/web/src/views/Pools/components/AutoEarningsBreakdown.tsx
@@ -1,5 +1,6 @@
 import { Text, Box } from '@pancakeswap/uikit'
 import { Pool } from '@pancakeswap/widgets-internal'
+import { BIG_ZERO } from '@pancakeswap/utils/bigNumber'
 
 import { useTranslation } from '@pancakeswap/localization'
 import { useVaultPoolByKey } from 'state/pools/hooks'
@@ -10,7 +11,7 @@ import { getCakeVaultEarnings } from '../helpers'
 
 interface AutoEarningsBreakdownProps {
   pool: Pool.DeserializedPool<Token>
-  account: string
+  account?: string
 }
 
 const AutoEarningsBreakdown: React.FC<React.PropsWithChildren<AutoEarningsBreakdownProps>> = ({ pool, account }) => {
@@ -22,18 +23,18 @@ const AutoEarningsBreakdown: React.FC<React.PropsWithChildren<AutoEarningsBreakd
   const { pricePerFullShare, userData } = useVaultPoolByKey(pool.vaultKey)
   const { autoCakeToDisplay, autoUsdToDisplay } = getCakeVaultEarnings(
     account,
-    userData.cakeAtLastUserAction,
-    userData.userShares,
-    pricePerFullShare,
-    earningTokenPrice,
+    userData?.cakeAtLastUserAction || BIG_ZERO,
+    userData?.userShares || BIG_ZERO,
+    pricePerFullShare || BIG_ZERO,
+    earningTokenPrice || 0,
     pool.vaultKey === VaultKey.CakeVault
       ? (userData as DeserializedLockedVaultUser).currentPerformanceFee
           .plus((userData as DeserializedLockedVaultUser).currentOverdueFee)
           .plus((userData as DeserializedLockedVaultUser).userBoostedShare)
-      : null,
+      : undefined,
   )
 
-  const lastActionInMs = userData.lastUserActionTime ? parseInt(userData.lastUserActionTime) * 1000 : 0
+  const lastActionInMs = userData?.lastUserActionTime ? parseInt(userData.lastUserActionTime) * 1000 : 0
   const hourDiffSinceLastAction = dayjs().diff(dayjs(lastActionInMs), 'hours')
   const earnedCakePerHour = hourDiffSinceLastAction ? autoCakeToDisplay / hourDiffSinceLastAction : 0
   const earnedUsdPerHour = hourDiffSinceLastAction ? autoUsdToDisplay / hourDiffSinceLastAction : 0