+++ apps/web/src/views/Pools/components/LockedPool/Modals/RoiCalculatorModalProvider.tsx
@@ -1,11 +1,11 @@
-import React, { useState, createContext } from 'react'
+import React, { useState, createContext, Dispatch, SetStateAction } from 'react'
 import { useTranslation } from '@pancakeswap/localization'
 import { usePool } from 'state/pools/hooks'
 import _toString from 'lodash/toString'
 
 import { VaultRoiCalculatorModal } from '../../Vault/VaultRoiCalculatorModal'
 
-export const RoiCalculatorModalContext = createContext(null)
+export const RoiCalculatorModalContext = createContext<Dispatch<SetStateAction<boolean>> | null>(null)
 
 const RoiCalculatorModalProvider: React.FC<
   React.PropsWithChildren<{ children: React.ReactNode; lockedAmount: string | number }>
@@ -17,7 +17,7 @@ const RoiCalculatorModalProvider: React.FC<
 
   const stakingTokenAddress = pool?.stakingToken?.address ? pool.stakingToken.address.toLowerCase() : null
 
-  if (showRoiCalculator) {
+  if (showRoiCalculator && pool) {
     return (
       <VaultRoiCalculatorModal
         pool={pool}