+++ apps/web/src/views/Swap/MMLinkPools/types.ts
@@ -101,7 +101,7 @@ export interface MMOrderBookTrade {
 export interface MMRfqTrade {
   rfq: RFQResponse['message'] | null | undefined
   trade: SmartRouterTrade<TradeType> | null
-  refreshRFQ: (() => void) | undefined
+  refreshRFQ: (() => void) | null
   quoteExpiry: number | null
   isLoading: boolean
   error?: Error